requirements.txt
h5py==2.9.0
pandas==0.25.0
numpy==1.16.4
dlib==19.18.0
scipy==1.3.1
scikit-image==0.15.0
scikit-learn==0.21.3
matplotlib==3.1.0
Pillow==6.2.0
imageio==2.5.0
imagesize==1.1.0
python-dateutil==2.8.0
pytest==5.0.1
librosa==0.7.0
numba==0.45.1
numexpr==2.6.9
pickleshare==0.7.5
audioread==2.1.8
youtube-dl==2019.9.28
face-recognition==1.2.3
face-recognition-models==0.3.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
tensorboard==2.0.0
tensorflow-gpu==2.0.0
tensorflow-estimator==2.0.0
ffmpeg==1.4

